#5 extract abstract BaseMethodProvider
